Separate sections for analysis and actions
